Refactoring classes.
